+++ src/hsm.cpp
@@ -18,14 +18,42 @@
  */
 #include "mococrw/hsm.h"
 
+#include <iomanip>
+#include <sstream>
+
+#include "libp11.h"
+
+#include "mococrw/key.h"
+
 namespace mococrw
 {
 using namespace openssl;
 
-HsmEngine::HsmEngine(const std::string &id, const std::string &modulePath, const std::string &pin)
-        : HSM(), _id(id), _modulePath(modulePath), _pin(pin)
+namespace
 {
-    // Fetch engine via ID.
+/**
+ * @brief Transforms given input to pct encoded string (see RFC 3986)
+ */
+std::string pctEncode(const std::vector<uint8_t> &bytes)
+{
+    if (bytes.size() == 0) {
+        return {};
+    }
+    std::stringstream result;
+    for (int byte : bytes) {
+        result << "%" << std::hex << std::setfill('0') << std::setw(2) << byte;
+    }
+    return result.str();
+}
+}  // namespace
+
+HsmEngine::HsmEngine(const std::string &id,
+                     const std::string &modulePath,
+                     const std::string &tokenLabel,
+                     const std::string &pin)
+        : _id(id), _modulePath(modulePath), _tokenLabel(tokenLabel), _pin(pin)
+{
+    // Fetch _engine via ID.
     _engine = _ENGINE_by_id(_id);
 
     _ENGINE_ctrl_cmd_string(_engine.get(), "MODULE_PATH", _modulePath);
@@ -35,14 +63,58 @@ HsmEngine::HsmEngine(const std::string &id, const std::string &modulePath, const
 
 HsmEngine::~HsmEngine() { _ENGINE_finish(_engine.get()); }
 
-openssl::SSL_EVP_PKEY_Ptr HsmEngine::loadPublicKey(const std::string &keyID)
+openssl::SSL_EVP_PKEY_Ptr HsmEngine::loadPublicKey(const std::string &keyLabel,
+                                                   const std::vector<uint8_t> &keyID) const
 {
-    return _ENGINE_load_public_key(_engine.get(), keyID);
+    auto keyIDPctEncoded = pctEncode(keyID);
+    std::string pkcs11URI =
+            "pkcs11:token=" + _tokenLabel + ";object=" + keyLabel + ";id=" + keyIDPctEncoded;
+    return _ENGINE_load_public_key(_engine.get(), pkcs11URI);
 }
 
-openssl::SSL_EVP_PKEY_Ptr HsmEngine::loadPrivateKey(const std::string &keyID)
+openssl::SSL_EVP_PKEY_Ptr HsmEngine::loadPrivateKey(const std::string &keyLabel,
+                                                    const std::vector<uint8_t> &keyID) const
 {
-    return _ENGINE_load_private_key(_engine.get(), keyID);
+    auto keyIDPctEncoded = pctEncode(keyID);
+    std::string pkcs11URI =
+            "pkcs11:token=" + _tokenLabel + ";object=" + keyLabel + ";id=" + keyIDPctEncoded;
+    return _ENGINE_load_private_key(_engine.get(), pkcs11URI);
 }
 
+openssl::SSL_EVP_PKEY_Ptr HsmEngine::generateKey(const RSASpec &spec,
+                                                 const std::string &keyLabel,
+                                                 const std::vector<uint8_t> &keyID)
+{
+    std::string keyIDHexString = utility::toHex(keyID);
+    PKCS11_RSA_KGEN pkcs11RSASpec;
+    pkcs11RSASpec.bits = spec.numberOfBits();
+    PKCS11_KGEN_ATTRS pkcs11RSAKeygen;
+    pkcs11RSAKeygen.type = EVP_PKEY_RSA;
+    pkcs11RSAKeygen.kgen.rsa = &pkcs11RSASpec;
+    pkcs11RSAKeygen.key_id = keyIDHexString.c_str();
+    pkcs11RSAKeygen.token_label = _tokenLabel.c_str();
+    pkcs11RSAKeygen.key_label = keyLabel.c_str();
+
+    _ENGINE_ctrl_cmd(_engine.get(), "KEYGEN", &pkcs11RSAKeygen);
+    return loadPrivateKey(keyLabel, keyID);
+}
+
+openssl::SSL_EVP_PKEY_Ptr HsmEngine::generateKey(const ECCSpec &spec,
+                                                 const std::string &keyLabel,
+                                                 const std::vector<uint8_t> &keyID)
+{
+    std::string curve = spec.curveName();
+    std::string keyIDHexString = utility::toHex(keyID);
+    PKCS11_EC_KGEN pkcs11ECCSpec;
+    pkcs11ECCSpec.curve = curve.c_str();
+    PKCS11_KGEN_ATTRS pkcs11ECCKeygen;
+    pkcs11ECCKeygen.type = EVP_PKEY_EC;
+    pkcs11ECCKeygen.kgen.ec = &pkcs11ECCSpec;
+    pkcs11ECCKeygen.key_id = keyIDHexString.c_str();
+    pkcs11ECCKeygen.token_label = _tokenLabel.c_str();
+    pkcs11ECCKeygen.key_label = keyLabel.c_str();
+
+    _ENGINE_ctrl_cmd(_engine.get(), "KEYGEN", &pkcs11ECCKeygen);
+    return loadPrivateKey(keyLabel, keyID);
+}
 }  // namespace mococrw