+++ src/key.cpp
@@ -107,9 +107,11 @@ AsymmetricPublicKey AsymmetricPublicKey::readPublicKeyFromPEM(const std::string 
 }
 
 #ifdef HSM_ENABLED
-AsymmetricPublicKey AsymmetricPublicKey::readPublicKeyFromHSM(HSM &hsm, const std::string &keyID)
+AsymmetricPublicKey AsymmetricPublicKey::readPublicKeyFromHSM(const HSM &hsm,
+                                                              const std::string &keyLabel,
+                                                              const std::vector<uint8_t> &keyID)
 {
-    auto key = hsm.loadPublicKey(keyID);
+    auto key = hsm.loadPublicKey(keyLabel, keyID);
     return AsymmetricPublicKey{std::move(key)};
 }
 #endif
@@ -186,10 +188,52 @@ AsymmetricKeypair AsymmetricKeypair::readPrivateKeyFromPEM(const std::string &pe
 }
 
 #ifdef HSM_ENABLED
-AsymmetricKeypair AsymmetricKeypair::readPrivateKeyFromHSM(HSM &hsm, const std::string &keyID)
+AsymmetricKeypair AsymmetricKeypair::readPrivateKeyFromHSM(const HSM &hsm,
+                                                           const std::string &keyLabel,
+                                                           const std::vector<uint8_t> &keyID)
 {
-    auto key = hsm.loadPrivateKey(keyID);
+    auto key = hsm.loadPrivateKey(keyLabel, keyID);
     return AsymmetricKeypair{std::move(key)};
+}
+
+AsymmetricKeypair AsymmetricKeypair::generateKeyOnHSM(HSM &hsm,
+                                                      const RSASpec &spec,
+                                                      const std::string &keyLabel,
+                                                      const std::vector<uint8_t> &keyID)
+{
+    // libp11 uses 128 byte buffer
+    if (keyID.size() >= 64) {
+        throw MoCOCrWException("Invalid keyID - key longer than 63 bytes");
+    }
+    try {
+        auto key = hsm.generateKey(spec, keyLabel, keyID);
+        return AsymmetricKeypair{std::move(key)};
+    } catch (const openssl::OpenSSLException &e) {
+        throw MoCOCrWException(
+                // wrong token-label? using unsupported ECC curve? HSM module implementation?
+                std::string("Key generation failed for unknown reason. OpenSSL error: ") +
+                e.what());
+    }
+}
+
+AsymmetricKeypair AsymmetricKeypair::generateKeyOnHSM(HSM &hsm,
+                                                      const ECCSpec &spec,
+                                                      const std::string &keyLabel,
+                                                      const std::vector<uint8_t> &keyID)
+{
+    // libp11 uses 128 byte buffer
+    if (keyID.size() >= 64) {
+        throw MoCOCrWException("Invalid keyID - key longer than 63 bytes");
+    }
+    try {
+        auto key = hsm.generateKey(spec, keyLabel, keyID);
+        return AsymmetricKeypair{std::move(key)};
+    } catch (const openssl::OpenSSLException &e) {
+        throw MoCOCrWException(
+                // wrong token-label? using unsupported ECC curve? HSM module implementation?
+                std::string("Key generation failed for unknown reason. OpenSSL error: ") +
+                e.what());
+    }
 }
 #endif