+++ src/mococrw/hsm.h
@@ -22,6 +22,8 @@
 
 namespace mococrw
 {
+class ECCSpec;
+class RSASpec;
 /**
  * The highest-level abstract class of a Hardware Security Module (HSM).
  *
@@ -43,25 +45,58 @@ protected:
     /**
      *  Loads public key from HSM.
      *
-     *  @param keyID The ID of the public key to load.
+     *  @param keyLabel String based identifer of a key on the token
+     *  @param keyID Vector of raw bytes that identifies a key on the token
      */
-    virtual openssl::SSL_EVP_PKEY_Ptr loadPublicKey(const std::string &keyID) = 0;
+    virtual openssl::SSL_EVP_PKEY_Ptr loadPublicKey(const std::string &keyLabel,
+                                                    const std::vector<uint8_t> &keyID) const = 0;
 
     /**
      * Loads private key from HSM.
      *
-     * @param keyID The ID of the private key to load.
+     * @param keyID Vector of raw bytes that identifies a key on the token
      */
-    virtual openssl::SSL_EVP_PKEY_Ptr loadPrivateKey(const std::string &keyID) = 0;
+    virtual openssl::SSL_EVP_PKEY_Ptr loadPrivateKey(const std::string &keyLabel,
+                                                     const std::vector<uint8_t> &keyID) const = 0;
+
+    /**
+     * @brief Generate a RSA key pair on the HSM
+     *
+     * @param spec The RSA specification @ref RSASpec
+     */
+    virtual openssl::SSL_EVP_PKEY_Ptr generateKey(const RSASpec &spec,
+                                                  const std::string &keyLabel,
+                                                  const std::vector<uint8_t> &keyID) = 0;
+
+    /**
+     * @brief Generate a ECC key pair on the HSM
+     *
+     * @param spec The ECC specification @ref ECCSpec
+     */
+    virtual openssl::SSL_EVP_PKEY_Ptr generateKey(const ECCSpec &spec,
+                                                  const std::string &keyLabel,
+                                                  const std::vector<uint8_t> &keyID) = 0;
 };
 
 /**
- * Abstract class of an HSMEngine that leverages OpenSSL's ENGINE_* API interface.
+ * Hsm handling that leverages OpenSSL's ENGINE_* API interface.
  */
 class HsmEngine : public HSM
 {
 public:
-    HsmEngine(const std::string &id, const std::string &modulePath, const std::string &pin);
+    /**
+     * @brief Constructor for an object that can manage keys on HSM using OpenSSL Engine
+     * @note Each HsmEngine object is associated with a specific token and a pin to login to that
+     * token
+     * @param id unique identifier for an OpenSSL engine
+     * @param modulePath path to HSM module i.e. softhsm
+     * @param tokenLabel label of the token where keys are managed
+     * @param pin pin to the mentioned token
+     */
+    HsmEngine(const std::string &id,
+              const std::string &modulePath,
+              const std::string &tokenLabel,
+              const std::string &pin);
     virtual ~HsmEngine();
 
 protected:
@@ -71,12 +106,24 @@ protected:
     const std::string _id;
     /** Path to Module. */
     const std::string _modulePath;
+    /** Token label used to uniquely identify a token on which objects reside */
+    const std::string _tokenLabel;
     /** Pin to access PKCS11 Engine. */
     const std::string _pin;
 
-    openssl::SSL_EVP_PKEY_Ptr loadPublicKey(const std::string &keyID) override;
+    openssl::SSL_EVP_PKEY_Ptr loadPublicKey(const std::string &keyLabel,
+                                            const std::vector<uint8_t> &keyID) const override;
 
-    openssl::SSL_EVP_PKEY_Ptr loadPrivateKey(const std::string &keyID) override;
+    openssl::SSL_EVP_PKEY_Ptr loadPrivateKey(const std::string &keyLabel,
+                                             const std::vector<uint8_t> &keyID) const override;
+
+    openssl::SSL_EVP_PKEY_Ptr generateKey(const RSASpec &spec,
+                                          const std::string &keyLabel,
+                                          const std::vector<uint8_t> &keyID) override;
+
+    openssl::SSL_EVP_PKEY_Ptr generateKey(const ECCSpec &spec,
+                                          const std::string &keyLabel,
+                                          const std::vector<uint8_t> &keyID) override;
 };
 
 }  // namespace mococrw